+++ src/components/ReusableConfirm/ReusableConfirm.tsx
@@ -1,8 +1,6 @@
 import { Button, Card, CardBody, CardFooter, Skeleton, Stack } from '@chakra-ui/react'
 import type { AssetId } from '@shapeshiftoss/caip'
-import { fromBaseUnit } from '@shapeshiftoss/utils'
 import type { PropsWithChildren, ReactNode } from 'react'
-import { useMemo } from 'react'
 import { useTranslate } from 'react-polyglot'
 
 import { Amount } from '@/components/Amount/Amount'
@@ -11,9 +9,6 @@ import { DialogHeader } from '@/components/Modal/components/DialogHeader'
 import { Row } from '@/components/Row/Row'
 import { SlideTransition } from '@/components/SlideTransition'
 import { RawText } from '@/components/Text'
-import { bn } from '@/lib/bignumber/bignumber'
-import { selectFeeAssetById, selectMarketDataByAssetIdUserCurrency } from '@/state/slices/selectors'
-import { useAppSelector } from '@/state/store'
 
 type ReusableConfirmProps = {
   assetId: AssetId
@@ -23,7 +18,7 @@ type ReusableConfirmProps = {
   fiatAmount: string
   confirmText: string
   feeAmountFiat: string | undefined
-  dustAmountCryptoBaseUnit?: string
+  dustAmountUserCurrency?: string
   isDisabled: boolean
   isLoading: boolean
   isError?: boolean
@@ -39,7 +34,7 @@ export const ReusableConfirm = ({
   cryptoSymbol,
   fiatAmount,
   feeAmountFiat,
-  dustAmountCryptoBaseUnit,
+  dustAmountUserCurrency,
   confirmText,
   isDisabled,
   isLoading,
@@ -50,22 +45,6 @@ export const ReusableConfirm = ({
   children,
 }: ReusableConfirmProps) => {
   const translate = useTranslate()
-
-  const feeAsset = useAppSelector(state => selectFeeAssetById(state, assetId))
-  const feeAssetMarketData = useAppSelector(state =>
-    selectMarketDataByAssetIdUserCurrency(state, feeAsset?.assetId ?? ''),
-  )
-
-  const dustAmountUserCurrency = useMemo(() => {
-    if (!dustAmountCryptoBaseUnit) return
-    if (!feeAsset) return
-
-    return bn(fromBaseUnit(dustAmountCryptoBaseUnit, feeAsset.precision))
-      .times(feeAssetMarketData.price)
-      .toFixed(2)
-  }, [dustAmountCryptoBaseUnit, feeAssetMarketData.price, feeAsset])
-
-  console.log({ dustAmountCryptoBaseUnit, dustAmountUserCurrency })
 
   return (
     <SlideTransition>
